refactor: remove P2 mocks after upstream merge
